requirements.txt
nats-py==2.9.0
influxdb-client[async]==1.50.0
aiohttp>=3.9
python-dateutil==2.9.0.post0
urllib3==2.6.3
